fix: cleanup a bit
