+++ wrappers/fedimint_clientd_rs/src/lib.rs
@@ -1,4 +1,2 @@
 mod client;
-pub mod options;
-pub mod types;
-pub use client::FedimintClient;
+pub use client::*;
