--- wrappers/fedimint_clientd_rs/src/options/mod.rs
@@ -1,7 +1,0 @@
-mod join;
-mod lightning;
-mod mint;
-
-pub use join::*;
-pub use lightning::*;
-pub use mint::*;
