--- wrappers/fedimint_clientd_rs/src/types/mod.rs
@@ -1,7 +1,0 @@
-mod requests;
-mod responses;
-mod values;
-
-pub use requests::*;
-pub use responses::*;
-pub use values::*;
